adding integer values

+++ db/seeds.rb
@@ -8,7 +8,7 @@ generation = chromosome.generations.create(iteration: 1)
 
 def add_value(index, value)
   name = @allele_names[index]
-  valuable = FloatValue.new(data: value)
+  valuable = IntegerValue.new(data: value)
   Value.create(organism: @organism, allele: Allele.find_by(name:), valuable:)
 end
